meaning explanation that starts from acts (!!)
This is in contrast to Martin-Löf's meaning explanations, which explain
verification and justification acts in terms of the existence of some
object; verification acts are the primary notion, and the analytic
judgments (equality, membership, etc.) are mere *reflections* on this
mathematical activity.

This explains the correpondence between (verification, justification)
and (canonical, noncanonical) satisfactorily, and it also shows what act
computation reflects---it is the recognition of a bar (or a cover).

@freebroccolo you may find these new slides interesting, starting with
the section entitled "The Creating Subject".